Move tests to a separate package
